fix file access problem on Windows
This is a problem in Windows seen in other libraries as well. Windows won' t delete a file because another process is still using it, according to itself, even though it isn't. It is just a windows timing issue. Hopefully mitigated now by explicitly closing memmap and file descriptor first.